终于debug完成了......

+++ MemoryManager/Interface/MemoryManager-1.cpp
@@ -1,7 +1,7 @@
 /*
  * @Date: 2022-03-24 19:43:00
  * @LastEditors: ShimaoZ
- * @LastEditTime: 2022-04-08 15:29:44
+ * @LastEditTime: 2022-04-08 17:34:13
  * @FilePath: \Operating-System\MemoryManager\Interface\MemoryManager-1.cpp
  */
 
@@ -9,8 +9,8 @@ class MemoryManager
 {
 
 public:
-    virtual int memoryAlloc(unsigned int pid, unsigned long length) = 0;
-    virtual bool memoryFree(unsigned int pid, unsigned long address, unsigned long length) = 0;
+    virtual int memoryAlloc(unsigned int pid, long long length) = 0;
+    virtual bool memoryFree(unsigned int pid, long long address, long long length) = 0;
     virtual bool freeAll(unsigned int pid) = 0;
-    virtual char accessMemory(unsigned int pid, unsigned long address) = 0; 
+    virtual char accessMemory(unsigned int pid, long long address) = 0; 
 };
