+++ MemoryManager/include/ProcessTableItem.h
@@ -1,7 +1,7 @@
 /*
  * @Date: 2022-04-01 15:07:54
  * @LastEditors: ShimaoZ
- * @LastEditTime: 2022-04-01 16:08:38
+ * @LastEditTime: 2022-04-08 17:34:05
  * @FilePath: \Operating-System\MemoryManager\include\ProcessTableItem.h
  */
 
@@ -11,7 +11,7 @@ class ProcessTableItem
 {
 private:
     unsigned int pid;
-    unsigned long pageCount;
+    long long pageCount;
     std::unordered_map<long, PageTableItem *> mPageTable;
 
 public:
